+++ src/main/java/Idea/Archieve/IdeaArchieve/domain/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
-package Idea.Archieve.IdeaArchieve.domain.chat.repository;
+package Idea.Archieve.IdeaArchieve.domain.comment.repository;
 
-import Idea.Archieve.IdeaArchieve.domain.chat.entity.Comment;
+import Idea.Archieve.IdeaArchieve.domain.comment.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
