+++ src/main/java/Idea/Archieve/IdeaArchieve/domain/post/Entity/Post.java
@@ -1,6 +1,6 @@
 package Idea.Archieve.IdeaArchieve.domain.post.Entity;
 
-import Idea.Archieve.IdeaArchieve.domain.chat.entity.Comment;
+import Idea.Archieve.IdeaArchieve.domain.comment.entity.Comment;
 import Idea.Archieve.IdeaArchieve.domain.member.Entity.Member;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
